Various

+++ code/client/cl_cin.c
@@ -494,7 +494,7 @@ int		spl;
 				index += 5;
 				break;
 		}
-	} while ( status[index] != NULL );
+	} while ( status[index] != NULL);
 }
 
 /******************************************************************************
@@ -1140,7 +1140,7 @@ redump:
 				cinTable[currentHandle].buf = 	cin.linbuf;
 			}
 			if (cinTable[currentHandle].numQuads == 0) {		// first frame
-				Com_Memcpy(cin.linbuf+cinTable[currentHandle].screenDelta, cin.linbuf, cinTable[currentHandle].samplesPerLine*cinTable[currentHandle].ysize);
+				Com_Memcpy(cin.linbuf+cinTable[currentHandle].screenDelta, cin.linbuf, (size_t)cinTable[currentHandle].samplesPerLine*cinTable[currentHandle].ysize);
 			}
 			cinTable[currentHandle].numQuads++;
 			cinTable[currentHandle].dirty = qtrue;
